stack name compose fix

+++ files/root/run.d/00-clear.passwd.sh
@@ -1,1 +1,3 @@
-yes | cp -f /etc/passwd.orig /etc/passwd
+if [ -f /etc/passwd.orig ]; then
+    yes | cp -f /etc/passwd.orig /etc/passwd
+fi
